refactor: address Z.ai review feedback
