+++ open-sse/services/rateLimitManager.ts
@@ -96,6 +96,7 @@ const PERSIST_DEBOUNCE_MS = 60_000; // Debounce persistence to every 60s max
 let initialized = false;
 
 let currentRequestQueueSettings: RequestQueueSettings = DEFAULT_RESILIENCE_SETTINGS.requestQueue;
+export const ZAI_WEB_REQUEST_QUEUE_MAX_WAIT_MS = 60_000;
 
 const limiterEffectiveSettings = new WeakMap<Bottleneck, Bottleneck.ConstructorOptions>();
 const preservedReplacementSettings = new Map<string, Bottleneck.ConstructorOptions>();
@@ -152,6 +153,15 @@ function resolveMinTime(override: number | undefined | null): number {
 // Resolve a maxConcurrent override. 0 or missing means "effectively infinite".
 function resolveMaxConcurrent(override: number | undefined | null): number {
   return typeof override === "number" && override > 0 ? override : EFFECTIVELY_INFINITE_CONCURRENCY;
+}
+
+export function resolveRequestQueueMaxWaitMs(
+  provider: string,
+  configuredMaxWaitMs: number = currentRequestQueueSettings.maxWaitMs
+): number {
+  return provider.trim().toLowerCase() === "zai-web"
+    ? Math.max(configuredMaxWaitMs, ZAI_WEB_REQUEST_QUEUE_MAX_WAIT_MS)
+    : configuredMaxWaitMs;
 }
 
 function buildLimiterDefaults() {
@@ -531,18 +541,19 @@ export async function withRateLimit(provider, connectionId, model, fn, signal = 
 
   // Proactive sliding-window fallback for header-less providers with a declared cap
   // (Fase 8.2). No-op unless PROVIDER_DEFAULT_RATE_LIMITS has an entry for `provider`.
+  const maxWaitMs = resolveRequestQueueMaxWaitMs(provider);
   await awaitProviderDefaultSlot(
     provider,
     connectionId,
     signal,
-    currentRequestQueueSettings.maxWaitMs
+    maxWaitMs
   );
 
   const limiter = getLimiter(provider, connectionId, model);
   // Bottleneck's `expiration` starts only after a job leaves QUEUED. The
   // legacy maxWaitMs setting therefore bounds limiter-managed execution; it
   // is not a queue-wait deadline.
-  const executionExpirationMs = currentRequestQueueSettings.maxWaitMs;
+  const executionExpirationMs = maxWaitMs;
   const scheduleOpts =
     executionExpirationMs && executionExpirationMs > 0 ? { expiration: executionExpirationMs } : {};
