+++ tests/unit/ratelimit-admission-control-6593.test.ts
@@ -1,12 +1,12 @@
 /**
- * #6593 — rate-limit queue admission control (maxQueueDepth) + bounded default.
+ * #6593 — rate-limit queue admission control (maxQueueDepth) + 15s default.
  *
  * RFC bundles 2 grounded changes to the local rate-limit request queue:
  *   1. `RequestQueueSettings.maxQueueDepth` — opt-in admission cap (default 0
  *      = disabled). When the queue already holds `maxQueueDepth` requests, a
  *      new request is fast-rejected with a typed `RATE_LIMIT_QUEUE_FULL`
  *      error instead of joining Bottleneck's queue.
- *   2. `DEFAULT_REQUEST_QUEUE_MAX_WAIT_MS` is bounded to 60 seconds.
+ *   2. `DEFAULT_REQUEST_QUEUE_MAX_WAIT_MS` lowered 120000 -> 15000.
  *
  * (Item #3 from the RFC, `bypassCompressionOnRateLimit`, has no matching
  * code path in this repo's compression pipeline — see the plan-file — and is
@@ -177,10 +177,16 @@ test("#6593 withRateLimit: default maxQueueDepth=0 preserves unbounded-queue beh
 
 // --- Default maxWaitMs ----------------------------------------------------
 
-test("#6593 DEFAULT_REQUEST_QUEUE_MAX_WAIT_MS is 60s absent RATE_LIMIT_MAX_WAIT_MS", () => {
+test("#6593 DEFAULT_REQUEST_QUEUE_MAX_WAIT_MS is 15s absent RATE_LIMIT_MAX_WAIT_MS", () => {
   assert.equal(process.env.RATE_LIMIT_MAX_WAIT_MS, undefined);
   assert.equal(resilienceSettings.DEFAULT_REQUEST_QUEUE_MAX_WAIT_MS, 15000);
   assert.equal(resilienceSettings.DEFAULT_RESILIENCE_SETTINGS.requestQueue.maxWaitMs, 15000);
+});
+
+test("#6593 zai-web receives a provider-scoped 60s scheduling budget", () => {
+  assert.equal(rateLimitManager.resolveRequestQueueMaxWaitMs("openai", 15_000), 15_000);
+  assert.equal(rateLimitManager.resolveRequestQueueMaxWaitMs("zai-web", 15_000), 60_000);
+  assert.equal(rateLimitManager.resolveRequestQueueMaxWaitMs("ZAI-WEB", 90_000), 90_000);
 });
 
 test("#6593 DEFAULT_REQUEST_QUEUE_MAX_DEPTH defaults to 0 (disabled) absent an env override", () => {
